+++ src/bluez/adapter.rs
@@ -1,5 +1,5 @@
 use super::peripheral::{Peripheral, PeripheralId};
-use crate::api::{Central, CentralEvent, CentralState, ScanFilter};
+use crate::api::{BDAddr, Central, CentralEvent, CentralState, ScanFilter};
 use crate::{Error, Result};
 use async_trait::async_trait;
 use bluez_async::{
@@ -124,6 +124,16 @@ impl Central for Adapter {
     async fn adapter_info(&self) -> Result<String> {
         let adapter_info = self.session.get_adapter_info(&self.adapter).await?;
         Ok(format!("{} ({})", adapter_info.id, adapter_info.modalias))
+    }
+
+    async fn adapter_address(&self) -> Result<Option<BDAddr>> {
+        let address: BDAddr = self
+            .session
+            .get_adapter_info(&self.adapter)
+            .await?
+            .mac_address
+            .into();
+        Ok((address != BDAddr::default()).then_some(address))
     }
 
     async fn adapter_state(&self) -> Result<CentralState> {